--- apps/server/src/middleware/Logger/index.ts
@@ -1,15 +1,0 @@
-import morgan from "morgan";
-import config from "../../config/index";
-
-const logType = (): string => {
-  switch (config.ENV) {
-    case "development":
-      return "dev";
-    default:
-      return "combined";
-  }
-};
-
-const Logger = morgan(logType());
-
-export default Logger;
